+++ src/index.js
@@ -4,3 +4,5 @@ export { default as useWillUnmount } from './useWillUnmount';
 export { default as useLifecycle } from './useLifecycle';
 export { default as useWindowResize } from './useWindowResize';
 export { default as useDebouncedCallback } from './useDebouncedCallback';
+export { default as useMouseHandler } from './useMouseHandler';
+export { default as useMouseState } from './useMouseState';